+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## FUTURE
+## v1.3.1
 
 * [FIXED] - `moduleLoader.fileIterator` did not support explicit file paths alongside directory paths even though public API claimed it as supported
 * [FIXED] - custom `Ajv` keyword `$toJSON` should support typeof string values which it should attemnt to parse